% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_image.R
\name{build_micromamba_image}
\alias{build_micromamba_image}
\title{Build Micromamba Container Image}
\usage{
build_micromamba_image(
  dockerfile_path = NULL,
  image_name = "umamba-dockerthis:latest"
)
}
\description{
Build Micromamba Container Image
}
