% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_build_image.R
\name{docker_build_image}
\alias{docker_build_image}
\title{Build Docker Container Image}
\usage{
docker_build_image(
  dockerfile_path,
  image_name,
  force = FALSE,
  platform_arg = NULL
)
}
\arguments{
\item{dockerfile_path}{Chracter. Path to Dockerfile.}

\item{image_name}{Character. Name of the Image to be created.}

\item{force}{Logical. Should image be removed before new build.
Default FALSE.}

\item{platform_arg}{Character. Platform string to be used, by default it
will use the one configured by Docker, usually Linux with the same
CPU architecture the same as the system,
import options are \code{c("linux/amd64", "linux/arm64", "linux/arm/v7")}.
Default NULL.}
}
\description{
Built Docker Image given a Dockerfile.
}
