% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{docker_client_cmd}
\alias{docker_client_cmd}
\alias{docker_list_images}
\alias{docker_list_containers}
\alias{docker_remove_image}
\alias{docker_remove_container}
\alias{docker_stop_container}
\alias{docker_start_container}
\title{Interact with Docker Engine Client CLI}
\usage{
docker_client_cmd(..., verbose = TRUE)

docker_list_images()

docker_list_containers()

docker_remove_image(image_name)

docker_remove_container(container_name)

docker_stop_container(container_name)

docker_start_container(container_name)
}
\arguments{
\item{...}{Character. Arguments to be passed to Docker CLI.}

\item{verbose}{Logical. Should output from CLI be printed. Default: \code{TRUE}.}

\item{image_name}{Character. Name or ID of the Image.}

\item{container_name}{Character. Name or ID of the Container.}
}
\description{
Functions for interaction with Docker Images and Containers
\itemize{
\item \code{docker_client_cmd()} is the main function to interact with the docker client.
\item \code{docker_list_images()} return a data frame with all tagged images
available for to user.
\item \code{docker_list_containers()} return a data  frame with all containers and
their state.
\item \code{docker_remove_image()} remove Docker image.
\item \code{docker_remove_container()} remove Docker container.
\item \code{docker_start_container()} start a stopped Docker container.
\item \code{docker_stop_container()} stop a running Docker container.
}
}
