% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_exec.R
\name{docker_exec}
\alias{docker_exec}
\title{Run Command on a Running Docker Container}
\usage{
docker_exec(cmd, ..., container_name, verbose = TRUE)
}
\arguments{
\item{cmd}{Character. Command to run.}

\item{...}{Character. Additional arguments passed to \code{cmd}.}

\item{container_name}{Name of Docker Container.}

\item{verbose}{Logical. Print execution output while running.
Default: \code{TRUE}.}
}
\description{
Run Command on a Running Docker Container
}
