% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{docker_remove_container}
\alias{docker_remove_container}
\title{Remove Docker Container}
\usage{
docker_remove_container(container_name)
}
\arguments{
\item{container_name}{Name or ID of the Container.}
}
\description{
Container needs to be stopped before removing
}
