% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{docker_remove_image}
\alias{docker_remove_image}
\title{Remove Docker Image}
\usage{
docker_remove_image(image_name)
}
\arguments{
\item{image_name}{Name or ID of the Image.}
}
\description{
Remove Docker Image
}
