% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_run.R
\name{docker_run}
\alias{docker_run}
\title{Create and Run Command on a New Docker Container}
\usage{
docker_run(
  cmd,
  ...,
  container_name,
  image_name,
  docker_args = NULL,
  mount_paths = NULL,
  verbose = TRUE
)
}
\arguments{
\item{cmd}{Character. Command to run.}

\item{...}{Character. Additional arguments passed to \code{cmd}.}

\item{container_name}{Name of Docker Container.}

\item{image_name}{Name of Container Image.}

\item{docker_args}{Character vector. Arguments passed to \verb{docker run} command.}

\item{mount_paths}{Character vector. Host paths to be mounted in container.}

\item{verbose}{Logical. Print execution output while running.
Default: \code{TRUE}.}
}
\description{
Start a new container to run command line tools inside Linux containers
using Docker Client CLI.
}
