% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerthis-package.R
\docType{package}
\name{dockerthis-package}
\alias{dockerthis}
\alias{dockerthis-package}
\title{dockerthis: Dockerthis: run any CLI tool on a Linux Container}
\description{
Dockerthis is an R package that simplifies the execution of command line tools within isolated Docker containers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luciorq/dockerthis}
  \item Report bugs at \url{https://github.com/luciorq/dockerthis/issues}
}

}
\author{
\strong{Maintainer}: Lucio Rezende Queiroz \email{luciorqueiroz@gmail.com} (\href{https://orcid.org/0000-0002-6090-1834}{ORCID})

}
\keyword{internal}
